move unit tests that require API access
